pipelines/requirements.txt
pandas
yfinance
psycopg2-binary
python-dotenv
xgboost
scikit-learn
joblib